MacBookPro Ready To Code
So I’ve got my new MacBookPro up and ready to code. Hopefully I can get
myself motivated to get some work done on the site. I need to get the
relationship status function working, and I believe I know a way to do
just that. I think I was trying to make it a lot harder than it needed
to be!