date formating
I’m starting to dive into the date formatting function built-in to php.
While this push only has a minor change to the show recipe page to
format the date of a comment differently, I’ll be adding a birthdate
area in the users database to store you birth date then after that I’ll
get a bio area figured out and push everything live as long as I don’t
encounter any major bugs in the site!