Site Displays Fully
I've gotten the site to display fully. Though I'm still not out of the woods yet, I've still got to get the MySQL connection working and figure out how to select, create, and update things. I've also addressed a few things in the about inc file and the allrecipe inc file. The allrecipe display script is still not working, but I'll address that soon.