Pulling the wrapper off
I've decided to pull the wrapper off the site. I'll probably use it for the content and sidebar div tags, but not for the footer and header/navigation. I'm also trying to tweak the navigation and header design: so much to do and so little time. But I'm having fun again and am determined to get the site fully functional by the middle to end of 2026. Then I'll look into making it generate passive income to supplement my lifestyle.